Added sudo indicator
If the user is currently in a valid sudo session (sudo currently does
not require a password), then the last part of the left prompt will use
a different color (orange). This gives an indication to the user that
sudo commands will not require a password at the moment. One example for
this might be that the user is reminded to run a `sudo -k` to terminate
the current sudo session before stepping away from the screen.