Allow to permanently disable proxy
Added XDG-compatible config directory: ~/.config/bash_it
This can also be accessed through the new ENV variable $BASH_IT_CONFIG.

Disabling the proxy will create a marker file at
$BASH_IT_CONFIG/proxy-disabled

If this file is present, the `init-proxy` function will not set
anything. If the file is not present, the proxy will be enabled.